+++ src/StreamStore.Benchmarking/Program.cs
@@ -1,0 +1,16 @@
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Running;
+
+namespace StreamStore.Benchmarking
+{
+    public static class Program
+    {
+        public static void Main(string[] args)
+        {
+            var config = DefaultConfig.Instance;
+
+            // Use this to select benchmarks from the console:
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args, config);
+        }
+    }
+}