+++ src/StreamStore.Contracts/Models/EventMetadataRecord.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+
 namespace StreamStore
 {
     public class EventMetadataRecord: IHasRevision
@@ -8,5 +10,16 @@ namespace StreamStore
         public DateTime Timestamp { get; set; }
 
         public int Revision { get; set; }
+    }
+
+    public class EventMetadataRecordCollection : RevisionedItemCollection<EventMetadataRecord>
+    {
+        public EventMetadataRecordCollection() : base()
+        {
+        }
+
+        public EventMetadataRecordCollection(IEnumerable<EventMetadataRecord> records) : base(records)
+        {
+        }
     }
 }