+++ src/StreamStore.Contracts/Models/EventRecord.cs
@@ -2,7 +2,7 @@
 {
     public class EventRecord : EventMetadataRecord
     {
-        public string? Data { get; set; }
+        public byte[]? Data { get; set; }
     }
 }