+++ src/StreamStore.S3.Tests/B2/B2S3FactoryTests.cs
@@ -19,7 +19,7 @@ namespace StreamStore.S3.Tests.B2
         public B2S3FactoryTests()
         {
             mockRepository = new MockRepository(MockBehavior.Strict);
-            factoryMock =new Mock<IStorageClientFactory>();
+            factoryMock = new Mock<IStorageClientFactory>();
             var fixture = new Fixture();
             settings = new B2StreamDatabaseSettingsBuilder()
                 .WithBucketId(fixture.Create<string>())
@@ -32,7 +32,7 @@ namespace StreamStore.S3.Tests.B2
         public void CreateClient_Should_CreateClientAndConnect()
         {
             // Arrange
-          
+
             var client = new Mock<IStorageClient>();
             factoryMock.Setup(m => m.Create()).Returns(client.Object);
             client.Setup(client => client.Connect(settings.Credentials!.AccessKeyId, settings.Credentials!.AccessKey));