+++ src/StreamStore.Testing/FixtureExtension.cs
@@ -6,11 +6,11 @@ namespace StreamStore.Testing
 {
     public static class FixtureExtension
     {
+        readonly static TypeRegistry registry = TypeRegistry.CreateAndInitialize();
 
         public static EventItem[] CreateEventItems(this Fixture fixture, int count)
         {
-            var serializer = new EventSerializer();
-
+            var serializer = new NewtonsoftEventSerializer(registry);
 
             var records =
                     fixture
@@ -22,22 +22,26 @@ namespace StreamStore.Testing
             return records;
         }
 
-
-        public static EventRecord[] CreateEventRecords(this Fixture fixture, int count)
+        public static EventRecord[] CreateEventRecords(this Fixture fixture, int initialRevision, int count)
         {
-            var serializer = new EventSerializer();
+            var serializer = new NewtonsoftEventSerializer(registry);
 
-            int revision = Revision.Zero;
+            var revision = initialRevision;
 
             var records =
                     fixture
                     .Build<EventRecord>()
-                    .With(x => x.Revision, () => ++revision)
+                    .With(x => x.Revision, () => revision++)
                     .With(x => x.Data, serializer.Serialize(fixture.Create<RootEvent>()))
                     .CreateMany(count)
                     .ToArray();
 
             return records;
+        }
+
+        public static EventRecord[] CreateEventRecords(this Fixture fixture,  int count)
+        {
+            return CreateEventRecords(fixture, 1, count);
         }
     }
 }