+++ src/StreamStore/ServiceCollectionExtension.cs
@@ -5,11 +5,12 @@ namespace StreamStore
 {
     public static class ServiceCollectionExtension
     {
-        public static IServiceCollection ConfigureStreamStore(this IServiceCollection services)
+        public static IServiceCollection ConfigureStreamStore(this IServiceCollection services, bool compression = true)
         {
-            services.AddSingleton<IStreamStore, StreamStore>();
-            services.AddSingleton<IEventSerializer, EventSerializer>();
-            return services;
+            return services
+                        .AddSingleton<IStreamStore, StreamStore>()
+                        .AddSingleton<ITypeRegistry>(services => TypeRegistry.CreateAndInitialize())
+                        .UserNewtonsoftJsonSerializer(compression);
         }
     }
 }